refactor(encoding): hoist common_prefix_len and level-1 constants into match_table::helpers
The Phase 1c split left `MatchGenerator::common_prefix_len` and the
level-1 hash-fill constants (`MIN_MATCH_LEN`, `FAST_HASH_FILL_STEP`,
`INCOMPRESSIBLE_SKIP_STEP`) hosted on `simple::MatchGenerator`, which
forced `match_table::helpers`, `dfast`, and `row` to import the Simple
matcher type just to call a free utility and forced `simple` to
re-import the constants from `match_generator` — closing a
`match_generator` ↔ `simple` reverse dependency.

Move both the constants and `common_prefix_len` into
`match_table::helpers` (the existing shared infra module) and switch
every matcher (`simple`, `dfast`, `row`) plus the `match_generator`
driver to depend on the shared module directly. No matcher imports
another matcher module anymore.

Behaviour is unchanged — `common_prefix_len` keeps the same
`#[inline(always)]` body that delegates to
`crate::encoding::fastpath::dispatch_common_prefix_len_ptr`, and the
constants keep their values. The remaining `match_generator::*`
imports on `dfast`/`row` are only for matcher-specific constants
(DFAST_*, ROW_*) that legitimately live next to the matcher driver
and are not shared.

Verified: clippy + 381 lib tests green (incl. level22 ratio gate).

Part of #111.
